seperate out sm64 code
